@@ -1,38 +1,26 @@
 
-import { Component, JsonObject, IHash, Config, Json } from "merapi";
-import { v4 as uuid } from "uuid";
 import { ICompile, IHelper, ITester } from "interfaces/main";
+import { Component, Config, IHash, JsonObject } from "merapi";
+import { v4 as uuid } from "uuid";
+
 const colors = require("colors");
-const inquirer = require("inquirer");
 const repl = require("repl");
 const util = require("util");
-const deasync = require("deasync");
 const Table = require("cli-table");
+const os = require("os");
+const fs = require("fs");
+const deasync = require("deasync");
 
 export default class Bot extends Component {
     constructor(private compile: ICompile, private helper: IHelper, private tester: ITester, private api: any) {
         super();
     }
 
-    public init(name: string, version: string, options: JsonObject) {
-        if (version) {
-            const versionRegex = /^\d+\.\d+\.\d+$/g;
-            if (!versionRegex.test(version)) {
-                console.log("Invalid version string");
-                console.log("Command kata init <botId> <botName> [version] is deprecated, use kata init <botName> [version] instead");
-                process.exit(0);
-            }
-        }
-
-        if (!version) {
-            version = "0.0.1";
-        }
-
+    public init(name: string, options: JsonObject) {
         const botDesc = {
             schema: "kata.ai/schema/kata-ml/1.0",
             name,
             desc: "My First Bot",
-            version: version || "0.0.1",
             flows: {
                 hello: {
                     fallback: true,
@@ -85,26 +73,18 @@ export default class Bot extends Component {
         console.log(`Initialized ${name} successfully`);
     }
 
-    public async versions(options: JsonObject) {
+    public async revisions(options: JsonObject) {
         try {
-            const botId = this.helper.getBotId();
-
-            if (!botId) {
-                throw new Error("BOT ID HAS NOT DEFINED");
-            }
-            const { data, response } = await this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdVersionsGet, botId);
-            if (data.versions) {
-                const history = data.versions;
-                console.log("Bot Versions : ");
-                history.data.forEach((hist: JsonObject) => {
-                    if (hist.tag) {
-                        console.log(`- ${hist.version} (${hist.tag})`);
-                    } else {
-                        console.log(`- ${hist.version}`);
-                    }
+            const projectId = this.getProject();
+            const { data, response } = await this.helper.toPromise(this.api.botApi, this.api.botApi.projectsProjectIdBotRevisionsGet, projectId);
+            const result = data;
+            if (result && result.data) {
+                console.log("Bot Revision : ");
+                result.data.forEach((hist: JsonObject) => {
+                    console.log(`- ${hist.revision}`);
                 });
             } else {
-                console.log("You must push at least 1 bot to acquire version");
+                console.log("You must push at least 1 bot to acquire revisions");
             }
         } catch (e) {
             if (e.code === "ENOENT") {
@@ -117,7 +97,7 @@ export default class Bot extends Component {
 
     public async test(file: string, options: JsonObject) {
         const testFiles = file ? [file] : this.helper.getFiles("./test", ".spec.yml");
-        const botId = this.helper.getBotId();
+        const botId = this.helper.getProjectId();
 
         if (!botId) {
             throw new Error("BOT ID HAS NOT DEFINED");
@@ -187,50 +167,8 @@ export default class Bot extends Component {
         }
     }
 
-    public async list(options: JsonObject) {
-        try {
-            const { data, response } = await this.helper.toPromise(this.api.botApi, this.api.botApi.botsGet, {});
-            const table = new Table({
-                head: ["Bot ID", "Bot Name", "Version", "Description"]
-                , colWidths: [20, 20, 10, 20]
-            });
-            data.items.forEach((bot: { id: string, name: string, version: string, desc: string }) => {
-                table.push([bot.id, bot.name, bot.version, bot.desc]);
-            });
-            console.log(table.toString());
-        } catch (e) {
-            console.log(this.helper.wrapError(e));
-        }
-    }
-
-    public async update(options: JsonObject) {
+    public async push(options: JsonObject) {
         const desc = this.helper.loadYaml("./bot.yml");
-
-        const oldVersion = desc.version;
-        let [major, minor, patch] = (desc.version as string).split(".").map((val: string) => parseInt(val));
-
-        switch (options.rev) {
-            case "major":
-                ++major;
-                minor = 0;
-                patch = 0;
-                break;
-            case "minor":
-                ++minor;
-                patch = 0;
-                break;
-            case "patch":
-                ++patch;
-                break;
-        }
-
-        if (major === undefined || minor === undefined || patch === undefined) {
-            major = major || 0;
-            minor = minor || 0;
-            patch = patch || 0;
-        }
-
-        desc.version = `${major}.${minor}.${patch}`;
         desc.tag = options.tag || null;
 
         let bot = Config.create(desc, { left: "${", right: "}" });
@@ -245,30 +183,14 @@ export default class Bot extends Component {
             return;
         }
 
-        if (!botDesc.id) {
-            const id = uuid();
-            botDesc.id = id;
-            desc.id = id;
-
-            try {
-                const result = await this.helper.toPromise(this.api.botApi, this.api.botApi.botsPost, botDesc);
-                console.log(`BOT CREATED SUCCESSFULLY WITH VERSION ${desc.version}`);
-            } catch (e) {
-                desc.version = oldVersion;
-
-                console.log(this.helper.wrapError(e));
-            }
-        } else {
-            try {
-                const result = await this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdPut, botDesc.id, botDesc, {});
-                desc.version = result.data.version;
-
-                console.log(`UPDATED BOT SUCCESSFULLY WITH VERSION ${desc.version}`);
-            } catch (e) {
-                desc.version = oldVersion;
-
-                console.log(this.helper.wrapError(e));
-            }
+        try {
+            const projectId = this.getProject();
+            botDesc.id = projectId;
+            const result = await this.helper.toPromise(this.api.botApi,
+                this.api.botApi.projectsProjectIdBotRevisionsPost, projectId, botDesc);
+            console.log(`Push Bot Success. Revision : ${result.data.revision.substring(0, 7)}`);
+        } catch (e) {
+            console.log(this.helper.wrapError(e));
         }
 
         this.helper.dumpYaml("./bot.yml", desc);
@@ -330,7 +252,7 @@ export default class Bot extends Component {
             return;
         }
 
-        const botId = this.helper.getBotId();
+        const botId = this.helper.getProjectId();
 
         try {
             const { data } = await this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdDelete, botId);
@@ -342,17 +264,15 @@ export default class Bot extends Component {
     }
 
     public console(options: JsonObject) {
-        let currentSession = (options.session ? options.session : uuid()) as string;
+        let projectId: string;
         let botDesc;
         try {
+            projectId = this.getProject() as string;
             botDesc = this.helper.loadYaml("./bot.yml");
         } catch (error) {
             console.log(this.helper.wrapError(error));
             return;
         }
-
-        const botId = botDesc.id;
-        const defaultDeploymentId = "f223c9e0-6ba1-434d-8313-a9f18ca364bd";
 
         const con = repl.start({
             prompt: botDesc.name + ">",
@@ -362,113 +282,64 @@ export default class Bot extends Component {
         });
 
         con.context.text = function text(str: string) {
+            let currentSession = this.getLocalSession();
             const message = {
                 type: "text",
                 content: str
             };
+
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
 
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-
-                return data;
+                return this.converse(projectId, body);
             } catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
 
         con.context.button = function button(op: JsonObject, obj: JsonObject = {}) {
+            let currentSession = this.getLocalSession();
             obj.op = op;
             const message = {
                 type: "data",
                 payload: obj
             };
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
 
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-
-                return data;
+                return this.converse(projectId, body);
             } catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
 
-        con.context.command = function button(command: string, obj: JsonObject = {}) {
+        con.context.command = function command(command: string, obj: JsonObject = {}) {
+            let currentSession = this.getLocalSession();
             const message = {
                 type: "command",
                 content: command,
                 payload: obj
             };
             const body = {
-                sessionId: currentSession,
+                session: currentSession,
                 message
             };
 
             try {
-                const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.botsBotIdConversePost, botId, body));
-
-                return data;
+                return this.converse(projectId, body);
             } catch (e) {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
 
-        con.context.current = function (session: string) {
-            if (arguments.length) {
-                currentSession = session;
-            } else {
-                return currentSession;
-            }
-        }.bind(this);
-
-        con.context.session = function session(name: string, update: JsonObject) {
-            try {
-                if (!arguments.length) {
-                    const res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, currentSession, "get"));
-
-                    return res.data;
-                } else if (arguments.length === 1) {
-                    const res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, name, "get"));
-
-                    return res.data;
-                } else {
-                    let res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, currentSession, "getOrCreate"));
-                    const session = res.data;
-                    res = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdPut, botId, defaultDeploymentId, session.id, update));
-
-                    return res.data;
-                }
-            } catch (e) {
-                return this.helper.wrapError(e);
-            }
-        }.bind(this);
-
-        con.context.clear = function clear(name: string) {
-            name = name || currentSession;
-
-            try {
-                const { data } = this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdGet, botId, defaultDeploymentId, name, "get"));
-                const session = { ...data };
-
-                if (session) {
-                    this.sync(this.helper.toPromise(this.api.sessionApi, this.api.sessionApi.botsBotIdDeploymentsDeploymentIdSessionsSessionIdDelete, botId, defaultDeploymentId, session.id));
-                }
-            } catch (e) {
-
-                if (e.status !== 400) {
-                    return;
-                }
-
-                return this.helper.wrapError(e);
-            }
-        }.bind(this);
+        con.context.current = () => this.getLocalSession();
+        con.context.clear = () => this.resetSession();
 
         con.context.clearCaches = function clearCaches(num: number = 20) {
             try {
@@ -479,6 +350,34 @@ export default class Bot extends Component {
                 return this.helper.wrapError(e);
             }
         }.bind(this);
+    }
+
+    private sync(promise: any) {
+        if (promise && typeof promise.then === "function") {
+            let done = false;
+            let error: Error = null;
+            let result;
+
+            promise.then((res: any) => {
+                done = true;
+                result = res;
+            }).catch((e: Error) => {
+                error = e;
+            });
+
+            deasync.loopWhile(() => {
+                return !done && !error;
+            });
+
+            if (error) {
+                throw error;
+            }
+
+            return result;
+        }
+
+
+        throw new Error("Sync only accept promises");
     }
 
     public async pull(name: string, version: string, options: JsonObject) {
@@ -515,31 +414,67 @@ export default class Bot extends Component {
         }
     }
 
-    private sync(promise: any) {
-        if (promise && typeof promise.then === "function") {
-            let done = false;
-            let error: Error = null;
-            let result;
+    private getProject() {
+        const projectId = this.helper.getProp("projectId")
 
-            promise.then((res: any) => {
-                done = true;
-                result = res;
-            }).catch((e: Error) => {
-                error = e;
-            });
-
-            deasync.loopWhile(() => {
-                return !done && !error;
-            });
-
-            if (error) {
-                throw error;
-            }
-
-            return result;
+        if (!projectId || projectId === "") {
+            throw Error("Error : You must specify a Project first, execute kata list-project to list your projects.");
         }
 
+        return projectId;
+    }
 
-        throw new Error("Sync only accept promises");
+    private converse(projectId: string, body: Object) {
+        const { data } = this.sync(this.helper.toPromise(this.api.botApi, this.api.botApi.projectsProjectIdBotConversePost, projectId, body)) as any;
+        const { session } = data;
+        this.setLocalSession(session);
+        return data;
+    }
+
+    private setLocalSession(session: Object) {
+        const jsonPath = `${os.homedir()}/.katasession`;
+        try {
+            if (session) {
+                fs.writeFileSync(jsonPath, JSON.stringify(session), "utf8");
+            }
+        } catch (error) {
+            console.log(this.helper.wrapError(`Error set local session : ${error.message}`));
+        }
+
+    }
+
+    private getLocalSession() {
+        const jsonPath = `${os.homedir()}/.katasession`;
+
+        if (fs.existsSync(jsonPath)) {
+            return JSON.parse(fs.readFileSync(jsonPath, "utf8"));
+        } else {
+            // default session
+            return {
+                "channel_id" : "console-channel",
+                "environment_id" : "console-environment",
+                "states" : {},
+                "contexes" : {},
+                "history" : [ ],
+                "current" : null,
+                "meta" : null,
+                "timestamp" : Date.now(),
+                "data" : {},
+                "created_at" : Date.now(),
+                "updated_at" : Date.now(),
+                "session_start" : Date.now(),
+                "session_id" : "test~from~console",
+                "id" : "test~from~console"
+            };
+        }
+    }
+
+    private resetSession() {
+        const jsonPath = `${os.homedir()}/.katasession`;
+
+        if (fs.existsSync(jsonPath)) {
+            fs.unlinkSync(jsonPath);
+            return true;
+        }
     }
 }
