@@ -9,20 +9,21 @@ class Api extends merapi_1.Component {
         this.apiClient = this.zaun.ApiClient.instance;
         let basePath = this.helper.getProp("zaunUrl") || "https://api.kata.ai";
         this.apiClient.basePath = basePath;
-        this.bearer = this.apiClient.authentications['Bearer'];
-        let currentLogin = this.helper.getProp("current_login") || "user";
-        let tokenObj = this.helper.getProp("token") || {};
+        this.bearer = this.apiClient.authentications.Bearer;
+        const currentLogin = this.helper.getProp("current_login") || "user";
+        const tokenObj = this.helper.getProp("token") || {};
         this.bearer.apiKey = `Bearer ${tokenObj[currentLogin]}`;
         this.botApi = new this.zaun.BotApi();
         this.authApi = new this.zaun.AuthApi();
         this.userApi = new this.zaun.UserApi();
         this.teamApi = new this.zaun.TeamApi();
         this.deploymentApi = new this.zaun.DeploymentApi();
-        this.draftApi = new this.zaun.DraftApi();
-        this.channelApi = new this.zaun.ChannelApi();
-        this.sessionApi = new this.zaun.SessionApi();
-        this.cachesApi = new this.zaun.CacheApi();
-        this.utilApi = new this.zaun.UtilApi();
+        this.projectApi = new zaun.ProjectApi();
+        // this.draftApi = new this.zaun.DraftApi();
+        // this.channelApi = new this.zaun.ChannelApi();
+        // this.sessionApi = new this.zaun.SessionApi();
+        // this.cachesApi = new this.zaun.CacheApi();
+        // this.utilApi = new this.zaun.UtilApi();
         this.nluApi = new this.zaun.NluApi();
     }
 }
