@@ -9,10 +9,9 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const merapi_1 = require("merapi");
-const colors = require("colors");
-const inquirer = require("inquirer");
 const Table = require("cli-table");
 const fs = require("fs");
+const yaml = require("js-yaml");
 const template = {
     default: {
         name: "kata",
@@ -113,71 +112,91 @@ class Nlu extends merapi_1.Component {
             }
         });
     }
+    pull() {
+        return __awaiter(this, void 0, void 0, function* () {
+            const projectId = this.helper.getProp("projectId");
+            try {
+                const { response: { body } } = yield this.helper.toPromise(this.api.projectApi, this.api.projectApi.projectsProjectIdNluGet, projectId);
+                const { name, lang, visibility, entities } = body;
+                const nluDesc = { name, lang, visibility, entities };
+                const nluYml = yaml.dump(nluDesc);
+                console.log("Writing to nlu.yml...");
+                fs.writeFileSync("nlu.yml", nluYml);
+            }
+            catch (error) {
+                console.log("Error: ", this.helper.wrapError(error));
+                return;
+            }
+        });
+    }
     push() {
         return __awaiter(this, void 0, void 0, function* () {
+            const projectId = this.helper.getProp("projectId");
             const nluDesc = this.helper.loadYaml("./nlu.yml");
             let nlu;
             let entities;
             try {
-                nlu = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameGet, nluDesc.name);
-                entities = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesGet, nluDesc.name);
+                const { response: { body } } = yield this.helper.toPromise(this.api.projectApi, this.api.projectApi.projectsProjectIdNluGet, projectId);
+                nlu = body;
             }
             catch (error) {
-                if (error.status === 400) {
-                    try {
-                        yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusPost, nluDesc);
-                        console.log(`NLU ${nluDesc.name} Created !`);
-                    }
-                    catch (error) {
-                        console.log(this.helper.wrapError(error));
-                    }
-                }
-                else {
-                    console.log(this.helper.wrapError(error));
-                }
+                console.log(this.helper.wrapError(error));
                 return;
             }
             try {
-                if (nlu && nlu.data) {
+                const { response: { body } } = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesGet, projectId, nluDesc.name);
+                entities = body;
+            }
+            catch (error) {
+                console.log(this.helper.wrapError(error));
+                return;
+            }
+            try {
+                if (nlu) {
                     let { lang, visibility } = nluDesc;
                     visibility = visibility || "private";
-                    yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNamePut, nluDesc.name, { lang, visibility });
-                    if (nluDesc.entities && entities.data) {
-                        const localDiff = this.helper.difference(nluDesc.entities, entities.data);
+                    // await this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNamePut,
+                    //     nluDesc.name, { lang, visibility });
+                    if (nluDesc.entities && entities) {
+                        const localDiff = this.helper.difference(nluDesc.entities, entities);
                         if (localDiff) {
                             for (const key in localDiff) {
-                                if (entities.data[key]) {
+                                if (entities[key]) {
                                     // Update remote entity
                                     if (!nluDesc.entities[key].inherit) {
-                                        yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesEntityNamePut, nluDesc.name, key, Object.assign({}, nluDesc.entities[key], { name: key }));
+                                        yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesEntityNamePut, projectId, nluDesc.name, key, Object.assign({}, nluDesc.entities[key], { name: key }));
                                     }
                                 }
                                 else {
                                     // Create new entity
-                                    yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesPost, nluDesc.name, Object.assign({}, nluDesc.entities[key], { name: key }));
+                                    yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesPost, projectId, nluDesc.name, Object.assign({}, nluDesc.entities[key], { name: key }));
                                 }
                             }
                         }
-                        const remoteDiff = this.helper.difference(entities.data, nluDesc.entities);
+                        const remoteDiff = this.helper.difference(entities, nluDesc.entities);
                         if (remoteDiff) {
                             for (const key in remoteDiff) {
                                 if (!nluDesc.entities[key]) {
                                     // delete remote entity
-                                    yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesEntityNameDelete, nluDesc.name, key);
+                                    yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesEntityNameDelete, projectId, nluDesc.name, key);
                                 }
                             }
                         }
                     }
-                    if (!nluDesc.entities && entities.data) {
-                        for (const key in entities.data) {
-                            // delete remote entity
-                            yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesEntityNameDelete, nluDesc.name, key);
+                    if (!nluDesc.entities && entities) {
+                        for (const key in entities) {
+                            if (entities[key]) {
+                                // delete remote entity
+                                yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesEntityNameDelete, projectId, nluDesc.name, key);
+                            }
                         }
                     }
-                    if (nluDesc.entities && !entities.data) {
+                    if (nluDesc.entities && !entities) {
                         for (const key in nluDesc.entities) {
-                            // create new entity
-                            yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameEntitiesPost, nluDesc.name, Object.assign({}, nluDesc.entities[key], { name: key }));
+                            if (nluDesc.entities[key]) {
+                                // create new entity
+                                yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameEntitiesPost, projectId, nluDesc.name, Object.assign({}, nluDesc.entities[key], { name: key }));
+                            }
                         }
                     }
                 }
@@ -190,8 +209,21 @@ class Nlu extends merapi_1.Component {
     }
     train(options) {
         return __awaiter(this, void 0, void 0, function* () {
+            const projectId = this.helper.getProp("projectId");
+            let nluDesc;
             try {
-                const nluDesc = this.helper.loadYaml("./nlu.yml");
+                nluDesc = this.helper.loadYaml("./nlu.yml");
+            }
+            catch (e) {
+                console.log("Missing nlu.yml");
+                return;
+            }
+            const nluName = nluDesc.name;
+            if (!nluName) {
+                console.log("Missing name in nlu.yml");
+                return;
+            }
+            try {
                 let opts = {};
                 if (options.file) {
                     console.log(`Training.. (input file: ${options.file})`);
@@ -205,7 +237,7 @@ class Nlu extends merapi_1.Component {
                         sentence: options.sentence
                     };
                 }
-                const trainResult = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameTrainPost, nluDesc.name, opts);
+                const trainResult = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNameTrainPost, projectId, nluName, opts);
                 console.log(`Success: ${trainResult.data.count} data trained !`);
             }
             catch (error) {
@@ -215,31 +247,41 @@ class Nlu extends merapi_1.Component {
     }
     predict(options) {
         return __awaiter(this, void 0, void 0, function* () {
+            const projectId = this.helper.getProp("projectId");
+            let nluDesc;
             try {
-                const nluDesc = !options.nlu ? this.helper.loadYaml("./nlu.yml") : { name: "" };
-                let nluName = nluDesc.name;
+                nluDesc = this.helper.loadYaml("./nlu.yml");
+            }
+            catch (e) {
+                console.log("Missing nlu.yml");
+                return;
+            }
+            const nluName = nluDesc.name;
+            if (!nluName) {
+                console.log("Missing name in nlu.yml");
+                return;
+            }
+            try {
                 let opts = {};
                 if (options.file) {
                     console.log(`Predict.. (input file: ${options.file})`);
                     opts = {
                         file: fs.createReadStream(options.file)
                     };
-                    nluName = options.nlu || nluName;
                 }
                 else if (options.sentence) {
                     console.log(`Predict.. (input: ${options.sentence})`);
                     opts = {
                         sentence: options.sentence
                     };
-                    nluName = options.nlu || nluName;
                 }
                 else {
                     throw new Error("Please input sentence or file to predict");
                 }
-                const predicResult = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNamePredictPost, nluName, opts);
+                const { response: { body } } = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNlusNluNamePredictPost, projectId, nluName, opts);
                 console.log(`Success, result : `);
                 let i = 0;
-                predicResult.response.body.result.forEach((x) => {
+                body.result.forEach((x) => {
                     console.log(`${++i}. Input: ${x.input}`);
                     console.log(`   Result: ${JSON.stringify(x.output)}`);
                 });
@@ -269,12 +311,14 @@ class Nlu extends merapi_1.Component {
             }
         });
     }
+    // deprecated soon
     listNlus(page, limit) {
         return __awaiter(this, void 0, void 0, function* () {
             try {
                 page = page || 1;
                 limit = limit || 10;
-                const nlus = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusGet, { page, limit });
+                const projectId = this.helper.getProp("projectId");
+                const nlus = yield this.helper.toPromise(this.api.projectApi, this.api.projectApi.projectsProjectIdNluGet, projectId, { page, limit });
                 if (nlus && nlus.data) {
                     const table = new Table({
                         head: ["Name", "Language", "Visibility"],
@@ -293,9 +337,10 @@ class Nlu extends merapi_1.Component {
     }
     snapshot() {
         return __awaiter(this, void 0, void 0, function* () {
+            const projectId = this.helper.getProp("projectId");
             try {
                 const nluDesc = this.helper.loadYaml("./nlu.yml");
-                const result = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.nlusNluNameSnapshotGet, nluDesc.name);
+                const result = yield this.helper.toPromise(this.api.nluApi, this.api.nluApi.projectsProjectIdNluSnapshotGet, projectId, nluDesc.name);
                 console.log(`Snapshot captured!`);
             }
             catch (error) {
